+++ CHANGELOG.md
@@ -6,6 +6,22 @@ This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a 
 
 ---
 
+## [3.4.0]
+
+### Added
+- Added `_twcOptions` to `tailwindClasses` function to allow for more flexible behavior (e.g., ignoring editor classes, disabling part debug classes, and indicating server-side rendering).
+
+### Fixed
+- SSR compatibility for `tailwindClasses` function, ensuring it works correctly in server-side rendering contexts.
+
+### Removed
+
+- Legacy `getTwPart`, `getTwDynamicPart`, `getTwClasses` functions (deprecated since 1.4.0).
+
+### Updated
+
+- Updated dependencies.
+
 ## [3.3.0]
 
 ### Changed
@@ -368,6 +384,7 @@ Co-authored with @piqusy
 - Initial release.
 
 [Unreleased]: https://github.com/infinum/eightshift-frontend-libs-tailwind/compare/master...HEAD
+[3.4.0]: https://github.com/infinum/eightshift-frontend-libs-tailwind/compare/3.4.0...3.4.0
 [3.3.0]: https://github.com/infinum/eightshift-frontend-libs-tailwind/compare/3.2.0...3.3.0
 [3.2.0]: https://github.com/infinum/eightshift-frontend-libs-tailwind/compare/3.1.2...3.2.0
 [3.1.2]: https://github.com/infinum/eightshift-frontend-libs-tailwind/compare/3.1.1...3.1.2